+++ src/app/View.ts
@@ -6,7 +6,7 @@ import createStoreFilters from './view-methods/store-page/create-store-filters';
 import createSortingSection from './view-methods/store-page/create-sorting-section';
 import { createProdDetailsContainer } from './view-methods/prod-detail-page/create-prod-details-container';
 import createCartSummary from './view-methods/cart-page/create-cart-summary';
-import { calculateImages } from './tools/Functions';
+import { calculateImages } from './tools/helpers';
 
 import {
     ProductDetails,
@@ -23,7 +23,7 @@ import {
 import createCartItem from './view-methods/cart-page/create-cart-item';
 import createCartContainer from './view-methods/cart-page/create-cart-container';
 import { CURRENCY_SYMBOL, SLIDER_MAX_ID, SLIDER_MIN_ID } from './constants/constants';
-import { checkSearchFocused } from './tools/Functions';
+import { checkSearchFocused } from './tools/helpers';
 import Cart from './Cart';
 
 class View {
@@ -87,9 +87,9 @@ class View {
     }
 
     renderImages(data: string[], title: string) {
-        const cont = document.querySelector('#details-images');
-        if (cont) {
-            cont.innerHTML += `
+        const imagesContainer = document.querySelector('#details-images');
+        if (imagesContainer) {
+            imagesContainer.innerHTML += `
                 ${data
                     .map((item) => {
                         return `<img src="${item}" alt="${title}">`;
